-refactoring the parser -moving parser to frontend
