+++ src/zwreec/frontend/mod.rs
@@ -1,4 +1,5 @@
 pub mod lexer;
+pub mod parser;
 
 pub fn lex<T: Iterator>(input :String)-> Vec<lexer::Token> {
     lexer::lex(input)
